--Initial Commit of https-bridge-PS-LV--
**Please ignore the copyright disclaimers that appear thoughout the project.
This originated as a personal project, and had been previously published to a
private GitHub repository under an MIT license.**

--Description--
This is a project that is intended to overcome the lack of functionality of
LabVIEW in regards to interacting with REST APIs. In order to do so this project
uses PowerShell, and opensource project managed by Microsoft and installed by
default sinces Windows 7. PowerShell is also already included in the environment
variables, meaning that the path is the same on all windows computers, which
removes the complications involving installing another .exe file and/or adding
a file to the Environment Path.
This project seeks to create a bridge from LabVIEW to PowerShell, allowing for
the use of PowerShell scripts in LabVIEW. This could potentially unlock much
needed methods in LabVIEW including HTTPS-PATCH, which is not supported natively
in LabVIEW